+++ CHANGES.md
@@ -4,6 +4,7 @@
   * Fixed broken `output_path` setting ([issue #169](https://github.com/Behat/Behat/issues/169))
   * Added shellbang to phar executable ([issue #167](https://github.com/Behat/Behat/issues/167))
   * Added feature title to progress exceptions ([issue #166](https://github.com/Behat/Behat/issues/166))
+  * Tuned failed formatter to print only failed examples in outline ([issue #154](https://github.com/Behat/Behat/issues/154))
 
 2.4.1 / 2012-06-26
 ==================